% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_cvma.R
\name{compare_control_diff}
\alias{compare_control_diff}
\title{compare_control for taking a difference in cross-validated maximal 
association measures.}
\usage{
compare_control_diff()
}
\description{
compare_control for taking a difference in cross-validated maximal 
association measures.
}
