% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_cvma.R
\name{compare_control_logratio}
\alias{compare_control_logratio}
\title{compare_control for taking a ratio of cross-validated maximal 
association measures, but computing CI and hypothesis test on log scale.}
\usage{
compare_control_logratio()
}
\description{
compare_control for taking a ratio of cross-validated maximal 
association measures, but computing CI and hypothesis test on log scale.
}
