% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_cvma.R
\name{compare_control_ratio}
\alias{compare_control_ratio}
\title{compare_control for taking a ratio of cross-validated maximal 
association measures.}
\usage{
compare_control_ratio()
}
\description{
compare_control for taking a ratio of cross-validated maximal 
association measures.
}
