% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_fn.R
\name{cv_risk_sl_auc}
\alias{cv_risk_sl_auc}
\title{Cross-validated area under the receiver operating characteristic curve
for predictions based on super learner}
\usage{
cv_risk_sl_auc(input, sl_control)
}
\arguments{
\item{input}{List where each entry corresponds to a validation fold. Each entry is a list
with entries: Y (univariate outcome for this validation fold), pred (matrix of predictions
from \code{learner} and columns correspond to different \code{learner}).}

\item{sl_control}{List of super learner control options.}
}
\description{
In general, the function passed to \code{sl_control$cv_risk} should expect a list
of outcomes and predictions in validation folds. The function should return a 
list with names cv_measure, ci_low, ci_high, and p_value. The output of this function 
is returned irrespective of the names of the list; however, the names are 
necessary for \code{print} methods to work properly.
}
\details{
In this case, the confidence intervals are computed using the \code{cvAUC::cvAUC.ci}
function from the \code{cvAUC} package. The p-value
is for the one-sided hypothesis test that cross-validated AUC equals 0.5 against
the alternative that it is greater than 0.5.
}
