% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_fn.R
\name{cv_risk_sl_nloglik}
\alias{cv_risk_sl_nloglik}
\title{Cross-validated negative log-likelihood of the super learner}
\usage{
cv_risk_sl_nloglik(input, sl_control)
}
\arguments{
\item{input}{List where each entry corresponds to a validation fold. Each entry is a list
with entries: Y (univariate outcome for this validation fold), pred (matrix of predictions
from \code{learner} and columns correspond to different \code{leaner}s).}

\item{sl_control}{List of super learner control options.}
}
\value{
A list with named entries cv_measure, ci_low, ci_high, and p_value. The list will
be returned by \code{max_assoc} irrespective of the named entries; however, the \code{print}
methods will only work if the function returns the above names.
}
\description{
In general, the function passed to \code{sl_control$cv_risk} should expect a list
of outcomes and predictions in validation folds. The function should return a 
list with names cv_measure, ci_low, ci_high, and p_value. The output of this function 
is returned irrespective of the names of the list; however, the names are 
necessary for \code{print} methods to work properly.
}
\details{
Confidence intervals are computed for negative log-likelihood. The p-value
is for the one-sided hypothesis test that cross-validated negative log-likelihood
is less than -log(0.5), which is what the value of negative log-likelihood if 
every observation were predicted to have Y = 1 with probability 0.5.
}
\examples{

# simulate data with proper format
input <- list(list(valid_folds=1, Y = rbinom(50,1,0.5),pred = runif(50,0,1)),
              list(valid_folds=2, Y = rbinom(50,1,0.5),pred = runif(50,0,1))) 
     
# alpha value                        
sl_control= list(alpha= 0.05)    

# get risk 
cv_risk <- cv_risk_sl_nloglik(input, sl_control)                         
}
