% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_fn.R
\name{cv_risk_y_auc}
\alias{cv_risk_y_auc}
\title{Cross-validated area under the receiver operating characteristic curve 
for evaluating maximum association.}
\usage{
cv_risk_y_auc(input, y_weight_control)
}
\arguments{
\item{input}{A list where each entry corresponds to a validation fold. Each entry is a list
with entries: Y (matrix of outcomes for this validation fold), 
pred (matrix of super learner predictions for each outcomes with columns corresponding to 
different outcomes).}

\item{y_weight_control}{Composite outcome weight control options.}
}
\value{
List with named components cv_measure (cross-validated AUC), ci_low (lower
100(1 - \code{y_weight_control$alpha})\% CI), ci_high (upper
100(1 - \code{y_weight_control$alpha})\% CI), p_value (p-value of test of null
hypothesis that cvAUC = 0.5).
}
\description{
In general, the function passed to \code{y_weight_control$cv_risk} should expect a list
of outcomes and predictions in validation folds, in addition to a list called
\code{y_weight} that contains the outcome weights (computed in the training sample)
corresponding to this validation fold and any other information needed by 
\code{y_weight_control$cv_risk} (e.g., anything needed to compute confidence 
intervals). The function should return a list with names cv_measure, ci_low,
ci_high, and p_value. The output of this function is returned irrespective of the
names of the list; however, the names are necessary for \code{print} methods to 
work properly.
}
\details{
In this case, the confidence intervals are computed using the \code{\link[cvAUC::cvAUC.ci]{cvAUC.ci}}
function from the \code{\link[cvAUC::cvAUC]{cvAUC}} package. The p-value
is for the one-sided hypothesis test that cross-validated AUC equals 0.5 against
the alternative that it is greater than 0.5.
}
\examples{

}
