% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_fn.R
\name{ensemble_linear}
\alias{ensemble_linear}
\title{Compute a linear ensemble of predictions}
\usage{
ensemble_linear(pred, weight)
}
\arguments{
\item{pred}{A matrix of predictions}

\item{weight}{A vector of ensemble weights (should have same length as number
of columns in \code{pred})}
}
\value{
A vector of ensembled values with length equal to the number of 
rows of \code{pred}.
}
\description{
In general, \code{sl_control$ensemble_fn} and \code{y_weight_control$ensemble_fn}
must accept named inputs, pred and weight, which are, respectively, a matrix and a vector. 
The matrix has columns corresponding to different learner fits (for \code{sl_control$ensemble_fn})
or to the different components of the multivariate outcome (for \code{y_weight_control$ensemble_fn}).
}
\details{
In this case, the function computes a linear combination.
}
