% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_fn.R
\name{ensemble_logit_linear}
\alias{ensemble_logit_linear}
\title{Compute a logit-linear ensemble of predictions}
\usage{
ensemble_logit_linear(pred, weight, l = 0, u = 1, trim = 0.001)
}
\arguments{
\item{pred}{A matrix of predictions}

\item{weight}{A vector of ensemble weights (should have same length as number
of columns in \code{pred})}

\item{l}{The lower scaling factor.}

\item{u}{The upper scaling factor}

\item{trim}{A value to trim logit computations to avoid numerical instabilities.}
}
\value{
A vector of ensembled values with length equal to the number of 
rows of \code{pred}.
}
\description{
In general, \code{sl_control$ensemble_fn} and \code{y_weight_control$ensemble_fn}
must accept named inputs, pred and weight, which are, respectively, a matrix and a vector. 
The matrix has columns corresponding to different learner fits (for \code{sl_control$ensemble_fn})
or to the different components of the multivariate outcome (for \code{y_weight_control$ensemble_fn}).
}
\details{
In this case, the function computes an ensemble of scaled (by \code{l} and \code{u})
\code{pred} on the logit scale and back-transforms to the original scale. The option
\code{trim} is used to avoid numeric instabilities.
}
