% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_fold_out}
\alias{get_fold_out}
\title{Helper function to format the validation folds used in
some of the \code{get_} functions}
\usage{
get_fold_out(x, split_Y, training_folds, valid_folds = NULL, V = NULL)
}
\description{
Helper function to format the validation folds used in
some of the \code{get_} functions
}
\keyword{internal}
