% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_fold_vector}
\alias{get_fold_vector}
\title{Helper function to get a vector of fold assignments (deprecated)}
\usage{
get_fold_vector(valid_y)
}
\description{
Helper function to get a vector of fold assignments (deprecated)
}
\keyword{internal}
