% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_formatted_sl}
\alias{get_formatted_sl}
\title{Get a super learner fit for a given outcome with more information.}
\usage{
get_formatted_sl(task, Y, V, all_fit_tasks, all_fits, folds, sl_control,
  return_learner_fits = TRUE, learners)
}
\arguments{
\item{task}{A named list identifying what outcome to use.}

\item{Y}{A matrix or data.frame of outcomes}

\item{V}{Number of outer folds of cross-validation (nested cross-validation
uses V-1 and V-2 folds), so must be at least four.}

\item{all_fit_tasks}{A list of all learner fitting tasks (quicker to search over
than \code{all_fits}).}

\item{all_fits}{A list of all learner fits (from \code{get_fit})}

\item{folds}{Vector identifying which fold observations fall into.}

\item{sl_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn, family. Available functions can be viewed with \code{sl_control_options()}. See
\code{?sl_control_options} for more on how users may supply their own functions.}

\item{return_learner_fits}{Should the \code{fit} component of the learners
be returned. Must be \code{TRUE} to obtain later predictions.}

\item{learners}{Super learner wrappers. See \code{\link{SuperLearner::listWrappers}}.}

\item{all_sl}{A list of all super learner fits (from \code{get_sl})}
}
\value{
Named list of super learner results.
}
\description{
This is called if \code{return_outer_sl = TRUE}, in which case more 
information on learner risks etc... is computed and returned.
}
