% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_pred_out}
\alias{get_pred_out}
\title{Helper function to get learner prediction matrices formatted
properly.}
\usage{
get_pred_out(fit_idx, learner_folds = NULL, sl_folds = NULL, all_fits,
  valid_folds = NULL, learners)
}
\description{
Helper function to get learner prediction matrices formatted
properly.
}
\keyword{internal}
