% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_risk}
\alias{get_risk}
\title{Get cross-validated risk of entire procedure (i.e., outer-most cross-validation layer)}
\usage{
get_risk(Y, V, all_fit_tasks, all_fits, all_weight, folds, all_sl, sl_control,
  y_weight_control, learners)
}
\arguments{
\item{Y}{A matrix or data.frame of outcomes}

\item{V}{Number of outer folds of cross-validation (nested cross-validation
uses V-1 and V-2 folds), so must be at least four.}

\item{all_fit_tasks}{A list of all learner fitting tasks (quicker to search over
than \code{all_fits}).}

\item{all_fits}{A list of all learner fits (from \code{get_fit})}

\item{all_weight}{A list of all outcome weights (from \code{get_y_weight})}

\item{folds}{Vector identifying which fold observations fall into.}

\item{all_sl}{A list of all super learner fits (from \code{get_sl})}

\item{sl_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn, family. Available functions can be viewed with \code{sl_control_options()}. See
\code{?sl_control_options} for more on how users may supply their own functions.}

\item{y_weight_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn. Available functions can be viewed with \code{y_weight_control_options()}. See
\code{?y_weight_control_options} for more on how users may supply their own functions.}

\item{learners}{Super learner wrappers. See \code{\link{SuperLearner::listWrappers}}.}
}
\value{
Numeric risk value
}
\description{
Get cross-validated risk of entire procedure (i.e., outer-most cross-validation layer)
}
