% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_risk_sl_input}
\alias{get_risk_sl_input}
\title{Create input list for \code{get_risk_sl}}
\usage{
get_risk_sl_input(split_Y, Yname, all_fits, learners, V, folds, all_sl,
  all_fit_tasks, all_weight, sl_control)
}
\arguments{
\item{split_Y}{The outcome matrix split by relevant validation folds.}

\item{Yname}{The names of the outcomes. Used to search \code{all_fits}
and \code{all_sl}.}

\item{all_fits}{List of all learner fits.}

\item{learners}{Vector of super learner vectors.}

\item{V}{Number of folds.}

\item{folds}{Cross-validation folds.}

\item{all_sl}{List of all super learner weight fits.}

\item{all_fit_tasks}{List of all learner fit tasks (faster to search over than
search over all_fits).}

\item{all_weight}{List of all outcome weight fits.}

\item{sl_control}{List of super learner controls.}
}
\value{
List with each entry a list with entries: valid_fold (the number of the
corresponding fold), Y (univariate outcome in this validation fold), pred (matrix
of outcome predictions for this validation fold from learners fit in 
training folds). \code{get_risk_sl} is only used to compute the cross-validated 
risk of individual outcomes and so is only used in the outer most cross-validation
layer.
}
\description{
Create input list for \code{get_risk_sl}
}
