% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_sl}
\alias{get_sl}
\title{Get super learner weights based on cross-validated learner fits}
\usage{
get_sl(task, Y, V, all_fit_tasks, all_fits, folds, sl_control, learners)
}
\arguments{
\item{task}{A named list identifying what training folds to use to 
obtain outcome weights.}

\item{Y}{A matrix or data.frame of outcomes}

\item{V}{Number of outer folds of cross-validation (nested cross-validation
uses V-1 and V-2 folds), so must be at least four.}

\item{all_fit_tasks}{A list of all learner fitting tasks (quicker to search over
than \code{all_fits}).}

\item{all_fits}{A list of all learner fits (from \code{get_fit})}

\item{folds}{Vector identifying which fold observations fall into.}

\item{sl_control}{A list with named entries ensemble_fn, optim_risk_fn, weight_fn,
cv_risk_fn, family. Available functions can be viewed with \code{sl_control_options()}. See
\code{?sl_control_options} for more on how users may supply their own functions.}

\item{learners}{Super learner wrappers. See \code{\link{SuperLearner::listWrappers}}.}
}
\value{
Named list identifying training folds used and the composite outcome weights.
}
\description{
Get super learner weights based on cross-validated learner fits
}
