% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_sl_input}
\alias{get_sl_input}
\title{Create input list for \code{get_sl}}
\usage{
get_sl_input(split_Y, Yname, valid_folds, V, all_fit_tasks, all_fits,
  learners)
}
\arguments{
\item{split_Y}{The outcome matrix split by relevant validation folds.}

\item{Yname}{The names of the outcomes. Used to search \code{all_fits}
and \code{all_sl}.}

\item{valid_folds}{If not \code{NULL}, identifies validation folds used in the
inner most super learner. Set to \code{NULL} when getting weights for outer
cross-validation folds (where all folds are used for training super learner).}

\item{V}{Number of folds.}

\item{all_fit_tasks}{List of all learner fit tasks (faster to search over than
search over all_fits).}

\item{all_fits}{List of all learner fits.}

\item{learners}{Vector of super learner vectors.}
}
\value{
List with each entry a list with entries: valid_fold (the number of the
corresponding fold), Y (univariate outcome in this validation fold), pred (matrix
of outcome predictions for this validation fold from learners fit in 
training folds). The length of the list varies according to what is input into 
\code{valid_folds} so that this function may be used both in the outer and inner
layers of cross-validation.
}
\description{
Create input list for \code{get_sl}
}
