% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_valid_pred_from_fit}
\alias{get_valid_pred_from_fit}
\title{Helper function to get validation fold predictions. Each
element of the \code{fit$pred} list will have a name
corresponding to the validation fold the predictions
were obtained on.}
\usage{
get_valid_pred_from_fit(fit, valid_folds)
}
\description{
Helper function to get validation fold predictions. Each
element of the \code{fit$pred} list will have a name
corresponding to the validation fold the predictions
were obtained on.
}
\keyword{internal}
