% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_valid_y}
\alias{get_valid_y}
\title{Helper function to get validation fold outcomes.}
\usage{
get_valid_y(split_Y, valid_folds)
}
\description{
Helper function to get validation fold outcomes.
}
\keyword{internal}
