% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_Y_out}
\alias{get_Y_out}
\title{Helper function to get outcomes in proper format needed
for some of the \code{get_} functions.}
\usage{
get_Y_out(x, split_Y, training_folds, V = NULL)
}
\description{
Helper function to get outcomes in proper format needed
for some of the \code{get_} functions.
}
\keyword{internal}
