% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_y_weight}
\alias{get_y_weight}
\title{Get outcome weights based on cross-validated super learner fits}
\usage{
get_y_weight(task, Y, V, Ynames, all_fits, all_sl, all_fit_tasks, sl_control,
  y_weight_control, folds, learners)
}
\arguments{
\item{task}{A named list identifying what training folds to use to 
obtain outcome weights.}

\item{Y}{A matrix or data.frame of outcomes}

\item{V}{Number of outer folds of cross-validation (nested cross-validation
uses V-1 and V-2 folds), so must be at least four.}

\item{Ynames}{Names of the columns of \code{Y}.}

\item{all_fits}{A list of all learner fits (from \code{get_fit})}

\item{all_sl}{A list of all super learner fits (from \code{get_sl})}

\item{all_fit_tasks}{A list of all learner fitting tasks (quicker to search over
than \code{all_fits}).}

\item{sl_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn, family. Available functions can be viewed with \code{sl_control_options()}. See
\code{?sl_control_options} for more on how users may supply their own functions.}

\item{y_weight_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn. Available functions can be viewed with \code{y_weight_control_options()}. See
\code{?y_weight_control_options} for more on how users may supply their own functions.}

\item{folds}{Vector identifying which fold observations fall into.}

\item{learners}{Super learner wrappers. See \code{\link{SuperLearner::listWrappers}}.}
}
\value{
Named list identifying training folds used and the composite outcome weights.
}
\description{
Get outcome weights based on cross-validated super learner fits
}
