% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_y_weight_input}
\alias{get_y_weight_input}
\title{Create input list for \code{get_y_weight}}
\usage{
get_y_weight_input(split_Y, Ynames, valid_folds, all_fits, all_sl,
  all_fit_tasks, sl_control, V, learners)
}
\arguments{
\item{split_Y}{The outcome matrix split by relevant validation folds.}

\item{Ynames}{The names of the outcomes. Used to search \code{all_fits}
and \code{all_sl}.}

\item{valid_folds}{If not \code{NULL}, identifies validation folds used in the
inner most super learner. Set to \code{NULL} when getting weights for outer
cross-validation folds (where all folds are used for training super learner).}

\item{all_fits}{List of all learner fits.}

\item{all_sl}{List of all super learner weight fits}

\item{all_fit_tasks}{List of all learner fit tasks (faster to search over than
search over all_fits).}

\item{sl_control}{List of super learner controls.}

\item{V}{Number of folds.}

\item{learners}{Vector of super learner vectors.}
}
\value{
List with each entry a list with entries: valid_fold (the number of the
corresponding fold), Y (the outcome matrix in this validation fold), pred (matrix
of outcome predictions for this validation fold from the super learner fit in 
training folds). The length of the list varies according to what is input into 
\code{valid_folds} so that this function may be used both in the outer and inner
layers of cross-validation.
}
\description{
Create input list for \code{get_y_weight}
}
