% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_options.R
\name{list_control_options}
\alias{list_control_options}
\title{List all control options in the \code{cvma} package.}
\usage{
list_control_options()
}
\description{
List all control options in the \code{cvma} package.
}
\examples{
list_control_options()
}
