% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lists.R
\name{make_outer_learner_task_list}
\alias{make_outer_learner_task_list}
\title{Helper function to make a task list for computing cv risks of learners}
\usage{
make_outer_learner_task_list(Ynames, V, learners)
}
\description{
Helper function to make a task list for computing cv risks of learners
}
\keyword{internal}
