% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lists.R
\name{make_outer_sl_task_list}
\alias{make_outer_sl_task_list}
\title{Helper function to make a task list for computing the outer super learner}
\usage{
make_outer_sl_task_list(Ynames, V)
}
\description{
Helper function to make a task list for computing the outer super learner
}
\keyword{internal}
