% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lists.R
\name{make_y_weight_task_list}
\alias{make_y_weight_task_list}
\title{Helper function to make a task list for computing outcome weights}
\usage{
make_y_weight_task_list(V)
}
\description{
Helper function to make a task list for computing outcome weights
}
