% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.cvma}
\alias{print.cvma}
\title{Print the output of \code{cvma}. Only prints 
the cross-validated summary measures. To print more
see \code{?summary.cvma}.}
\usage{
\method{print}{cvma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cvma}}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of \code{cvma}. Only prints 
the cross-validated summary measures. To print more
see \code{?summary.cvma}.
}
