% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fits.R
\name{search_fits_all_y}
\alias{search_fits_all_y}
\title{Helper function to search fits for particular training_folds.}
\usage{
search_fits_all_y(l, training_folds)
}
\description{
Helper function to search fits for particular training_folds.
}
\keyword{internal}
