% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fits.R
\name{search_fits_for_learner}
\alias{search_fits_for_learner}
\title{Helper function to search fits for particular outcomes (or single outcomes)
and training_folds.}
\usage{
search_fits_for_learner(training_folds, fits, learner, y = NULL)
}
\description{
Helper function to search fits for particular outcomes (or single outcomes)
and training_folds.
}
\keyword{internal}
