% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fits.R
\name{search_fits_one_learner}
\alias{search_fits_one_learner}
\title{Helper function to search fits for particular outcomes and training_folds.}
\usage{
search_fits_one_learner(l, Yname, training_folds, learner)
}
\description{
Helper function to search fits for particular outcomes and training_folds.
}
\keyword{internal}
