% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fits.R
\name{search_fits_one_y}
\alias{search_fits_one_y}
\title{Helper function to search fits for particular outcomes and training_folds.}
\usage{
search_fits_one_y(l, Yname, training_folds)
}
\description{
Helper function to search fits for particular outcomes and training_folds.
}
\keyword{internal}
