% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.cvma}
\alias{summary.cvma}
\title{Summarize results of \code{cvma} fit. Can be used to summarize
the weights obtained for each outcome (in both outer and inner
layers of cross-validation), the super learner fit for each outcome
(with learners fit on all the data), or the cross-validated associations
with each outcome.}
\usage{
\method{summary}{cvma}(object, aspect = "outcomes", ...)
}
\arguments{
\item{object}{An object of class \code{cvma}}

\item{aspect}{Can be \code{"weights"}, \code{"superlearner"},
\code{"outcomes"}, or \code{"learners"}}

\item{...}{Other options (not currently used)}
}
\value{
A list summarizing relevant aspects of the \code{cvma} object
}
\description{
Summarize results of \code{cvma} fit. Can be used to summarize
the weights obtained for each outcome (in both outer and inner
layers of cross-validation), the super learner fit for each outcome
(with learners fit on all the data), or the cross-validated associations
with each outcome.
}
