% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvma.R
\name{cvma}
\alias{cvma}
\title{Cross-validated maximal association measures}
\usage{
cvma(Y, X, V = 5, learners, sl_control = list(ensemble_fn =
  "ensemble_linear", optim_risk_fn = "optim_risk_sl_se", weight_fn =
  "weight_sl_convex", cv_risk_fn = "cv_risk_sl_r2", family = gaussian(), alpha =
  0.05), y_weight_control = list(ensemble_fn = "ensemble_linear", weight_fn =
  "weight_y_convex", optim_risk_fn = "optim_risk_y_r2", cv_risk_fn =
  "cv_risk_y_r2", alpha = 0.05), return_outer_weight = TRUE,
  return_outer_sl = TRUE, return_all_y = TRUE, scale = FALSE)
}
\arguments{
\item{Y}{A matrix or data.frame of outcomes}

\item{X}{A matrix or data.frame of predictors}

\item{V}{Number of outer folds of cross-validation (nested cross-validation
uses V-1 and V-2 folds), so must be at least four.}

\item{learners}{Super learner wrappers. See \code{SuperLearner::listWrappers}.}

\item{sl_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn, family. Available functions can be viewed with \code{sl_control_options()}. See
\code{?sl_control_options} for more on how users may supply their own functions.}

\item{y_weight_control}{A list with named entries ensemble.fn, optim_risk_fn, weight_fn,
cv_risk_fn. Available functions can be viewed with \code{y_weight_control_options()}. See
\code{?y_weight_control_options} for more on how users may supply their own functions.}

\item{return_outer_weight}{Whether to return estimate of outcome
weights the outer most cross-validation layer (i.e, based on V - 1 cross-validated 
super learner risks).}

\item{return_outer_sl}{Whether to return the super learner for the outer most cross-validation
layer (i.e., V-fold super learner).}

\item{return_all_y}{Whether to return cross-validated performance measures for each
column of \code{Y}.}

\item{scale}{Standardize each outcome to be mean zero with standard deviation 1.}
}
\value{
TO DO : Add return values
}
\description{
TO DO: Add a description here.
}
\details{
TO DO: Figure out how future works (e.g., can plan() be specified internally
or externally?)
}
\examples{
set.seed(1234)
library(SuperLearner)
library(future)
X <- data.frame(x1=runif(n=100,0,5), x2=runif(n=100,0,5))
Y1 <- rnorm(100, X$x1 + X$x2, 1)
Y2 <- rnorm(100, X$x1 + X$x2, 3)
Y <- data.frame(Y1 = Y1, Y2 = Y2)
fit <- cvma(Y = Y, X = X, V = 5, 
                learners = c("SL.glm","SL.mean"))


}
\seealso{
predict method
}
