% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_fn.R
\name{weight_sl_convex}
\alias{weight_sl_convex}
\title{Convex ensemble weights for super learner}
\usage{
weight_sl_convex(input, sl_control)
}
\arguments{
\item{input}{A list where each entry corresponds to a validation fold. Each entry is a list
with entries: Y (univariate outcome for this validation fold), pred (matrix of predictions
from \code{learner} and columns correspond to different \code{learner}).}

\item{sl_control}{Super learner control options.}
}
\value{
Numeric vector giving convex weights for super learner.
}
\description{
In general, the function passed to \code{sl_control$weight_fn} should expect a list of 
named lists of outcomes (Y), predictions (pred) in validation folds. Typically,
this function is used to maximize \code{sl_control$optim_risk_fn} over 
weights. The function should return a named list. One of the names in the list should
be \code{weight}, which is the optimized weights. Other entries in the return list
are passed on to \code{sl_control$cv_risk_fn} (e.g., things needed to compute
cross-validated measure of association, though none are present for this particular 
function).
}
\details{
In this case, the function uses \code{Rsolnp::solnp} to minimize 
\code{sl_control$optim_risk_fn} over convex weights.
}
\examples{


}
